A new SLang function: mc->action("EditFile") or mc->action(507) – initiating of given CK_ action.
The return value is TRUE if any of the called callbacks (2 at this point)
returned MSG_HANDLED. First edit_dialog_callback() is tried and if it doesn't
handle the action, then the action is passed to edit_execute_key_command().
If this function also returns MSG_NOT_HANDLED, then FALSE is returned.